Cleanup vaccel ops file organization
Better organization of operations code in a per type of operation files.
This will make "easier" the process of including a new operation in the
vAccel API.
