+++ plugins/virtio/operations.c
@@ -1,4 +1,5 @@
 #include <vaccel.h>
+#include <ops/vaccel_ops.h>
 #include <stddef.h>
 #include <sys/ioctl.h>
 #include <accel.h>
