+++ src/ops/image.h
@@ -1,0 +1,21 @@
+#ifndef __VACCEL_IMAGE_OPS_H__
+#define __VACCEL_IMAGE_OPS_H__
+
+#include <stddef.h>
+
+struct vaccel_session;
+
+int vaccel_image_classification(struct vaccel_session *sess, const void *img,
+		unsigned char *out_text, unsigned char *out_imgname,
+		size_t len_img, size_t len_out_text, size_t len_out_imgname);
+
+int vaccel_image_detection(struct vaccel_session *sess, const void *img,
+		const unsigned char *out_imgname, size_t len_img,
+		size_t len_out_imgname);
+
+int vaccel_image_segmentation(struct vaccel_session *sess, const void *img,
+		const unsigned char *out_imgname, size_t len_img,
+		size_t len_out_imgname);
+
+
+#endif /* __VACCEL_IMAGE_OPS_H__ */
