+++ src/ops/noop.h
@@ -1,0 +1,8 @@
+#ifndef __VACCEL_NOOP_H__
+#define __VACCEL_NOOP_H__
+
+struct vaccel_session;
+
+int vaccel_noop(struct vaccel_session *sess);
+
+#endif /* __VACCEL_NOOP_H__ */
