+++ src/vaccel.h
@@ -5,7 +5,9 @@
 #include <stddef.h>
 
 #include "common.h"
-#include "vaccel_ops.h"
+
+#include "ops/image.h"
+#include "ops/noop.h"
 
 #ifdef __cplusplus
 extern "C" {
